+++ src/App.js
@@ -11,6 +11,7 @@ import $ from 'jquery';
 import People from './components/People';
 import SaveButton from './components/SaveButton';
 import NicetyDisplay from './components/NicetyDisplay';
+import Admin from './components/Admin';
 
 // var updated_niceties_spinlock = false;
 // var updated_niceties = new Set();
@@ -21,68 +22,8 @@ if (localStorage.getItem("saved") === null || localStorage.getItem("saved") === 
 }
 
 
-
-
-
-var Admin = React.createClass({
-    loadAllNiceties: function(callback) {
-        $.ajax({
-            url: this.props.admin_edit_api,
-            dataType: 'json',
-            cache: false,
-            success: function(data) {
-                callback(data);
-            },
-            error: function(xhr, status, err) {
-                console.error(this.props.people, status, err.toString());
-            }.bind(this)
-        });
-    },
-    getInitialState: function() {
-        return {
-            niceties: []
-        }
-    },
-    componentDidMount: function() {
-        this.loadAllNiceties(function (data) {
-            this.setState({niceties: data});
-        }.bind(this));
-    },
-    render: function() {
-        return (
-            <div>
-                {this.state.niceties.map((person) => {
-                    let noTextCheck = false;
-                    person.niceties.forEach((nicety) => {
-                        if (nicety.text !== '' && nicety.text !== null) {
-                            noTextCheck = true;
-                        }
-                    });
-                    if (noTextCheck) {
-                        return (
-                            <div>
-                            <h2>To {person.to_name}</h2>
-                                <br />
-                                {person.niceties.map((nicety) => {
-                                    return (
-                                            <AdminNicety nicety={nicety} target_id={person.to_id}
-                                                         admin_edit_api={this.props.admin_edit_api}/>
-                                    );
-                                })}
-                            <hr />
-                            </div>
-                        );
-                    } else {
-                        return null;
-                    }
-                })}
-            </div>
-        );
-    }
-});
-
 // abc
-var AdminNicety = React.createClass({
+export var AdminNicety = React.createClass({
     getInitialState: function() {
         return {
             text: this.props.nicety.text,
